refactor(modules): 更新 modules 名称与引入路径
