+++ src/modules/courseTable/components/courseTable/index.tsx
