+++ src/modules/courseTable/components/courseTable/type.ts
