+++ src/modules/courseTable/components/weekSelector/index.tsx
