+++ src/modules/mainPage/components/courseTree/index.tsx
