+++ src/modules/mainPage/components/grid/type.ts
