+++ src/modules/setting/index.tsx
@@ -3,7 +3,7 @@ import { FlatList } from 'react-native';
 import ThemeBasedView from '@/components/view';
 
 import { SettingItems } from '@/constants/settingItem';
-import SettingItem from '@/module/setting/components/settingItem';
+import SettingItem from '@/modules/setting/components/settingItem';
 
 export default function SettingPage() {
   return (
