+++ src/utils/index.ts
@@ -1,6 +1,7 @@
 export { default as EventBus } from './eventBus';
 export { default as fetchUpdate } from './fetchUpdate';
 export { default as generateStyleSheet } from './generateStyleSheet';
+export { default as handleCopy } from './handleCopy';
 export { default as handleOpenURL } from './handleOpenURL';
 export { default as keyGenerator } from './keyGenerator';
 export { default as percent2px } from './percent2px';
